DEP-494: Switching bundler-smrt-services to tomcat_8.5.20
SWitching bundler-smrt-services to tomcat_8.5.20, mostly trying to minimize
the changes, but making some steps to make it easier to allow subsequent
tomcat upgrades.

This will use the tomcat-pbtarbal_8.5.20 tarball as input, which is a
modification of the stock tomcat tarball_8.5.20 to make the pacbio specific
changes for configuration and hardening that will be shared between the
smrtview and smrtlink instances of tomcat.  The changes to the stock tarball
include:

  - removing all group/other permissions on the entire tomcat directory
  - disabling use of the shutdown port in the stop script
  - reducing verbosity in startin up the tomcat server
  - disabling the shutport
  - disabling the AJP port

For smrtflow, changes include:

  - using a tomcat_current link for anything after the tarball has been
    extracted (should make it easier for future tomcat upgrades)
  - allowing tomcat and ivy cache to be passed in to bamboo_build.sh (should
    make it easier to specify the correct input for builds on different
    branches)
  - upgrading to jdk_8u144 (from jdk_8u71)
  - using the tomcat-pbtarball_8.5.20 tarball
  - better detection the extracted root directory in the tarball

There are still other improvementst that could/should be made to make it
easier to upgrade tomcat and support builds on different branches, but this
will get us a good look at what to expects for the tomcat upgrades for jaguar.