DEP-524: bumping chemistry to 5.0.1.14474, permissions fixed
Updating to the new version of the released chemistry (5.0.1.14474) that
has the permissions fixed.

According to @wmcnamara, the permissions get modified when downloaded from the
update server by smrtlink, but not when added directly to the build.  The
5.0.1.14474 version and future chemistry bundle vesions will have normal
permission, so this should not be a problem going forward.

Note there will probably be a changes to some 5.1.0 chemistry for the J
production release.