DEP-397: dbctrl: force encoding and locale at database initdb/createdb time
Force the encoding to UTF-8 and locale to en_US.UTF-8 at the database
initdb and createdb time in dbctrl.  For fresh installs, this should force the
databases to be set up with en_US.UTF-8, regardless of how the system or the
user specified the current locale.

This is mostly a safety measure since we've had reports of LATIN* encodings in
the database, even if the locale was set to en_US.UTF-8.

This does not affect the encoding/locale of previously created databases (as
in the case of an upgrade).  We will need to detect those and deal with them
in a different way.

Note that this does not replace setting of the locale environment via LC_ALL,
which we still intend to do for all the smrtlink entry points.

Also note that the postgres, glibc and other packages are forgiving and not
necessarily consistent in how encodings are specified (e.g. case may not be
sensitive and hyphens may be optional, so en_US.UTF-8 may be equivalent to
en_US.utf8).  For the createdb and initdb command lines modified here, we use
"UTF-8" and "en_US.UTF-8" to try to be consistent with how we refer to the
locale in other areas.  But postgres via psql will report the encoding as
"UTF8" and the locale as "en_US.UTF-8":

    postgres=# \l
                                          List of databases
        Name    |     Owner     | Encoding |   Collate   |    Ctype    |        Access privileges
    ------------+---------------+----------+-------------+-------------+---------------------------------
     postgres   | smrtlink_user | UTF8     | en_US.UTF-8 | en_US.UTF-8 |
     smrtlinkdb | smrtlink_user | UTF8     | en_US.UTF-8 | en_US.UTF-8 |
     template0  | smrtlink_user | UTF8     | en_US.UTF-8 | en_US.UTF-8 | =c/smrtlink_user               +
                |               |          |             |             | smrtlink_user=CTc/smrtlink_user
     template1  | smrtlink_user | UTF8     | en_US.UTF-8 | en_US.UTF-8 | =c/smrtlink_user               +
                |               |          |             |             | smrtlink_user=CTc/smrtlink_user
     wso2am     | smrtlink_user | UTF8     | en_US.UTF-8 | en_US.UTF-8 |
    (5 rows)