DEP-581: double number of retries (timeout length) in smrtlink-gui get-status
Increase the number of retries from 30 to 60 when waiting for smrtlink-gui
(tomcat) to start up, increasing the timeout from 155 seconds to 305 seconds.

For the smrtlink-gui case, we are seeing that some installations on local
disks
will take slightly longer than the allotted timeout to start up.  We may see a
greater frequency of that on nfs.  Note, we see the premature smrtlink-gui
timeouts in about 1 out of 200 starts or so (on local disks).

Doubling the timeout should address the issue for most cases.  And since we do
not often see unsuccessful tomcat startups, the penalty of waiting longer in
the timeout case should not be a big deal.

Note, now specifying the smrtlink-services and smrtlink-gui retries
independently, since there does not seem to be a need to increast the smrtlink
timeout.