DEP-939: handle nulls in wso2carbon.pid to suppress bash-4.4 warnings
Handle nulls in wso2carbon.pid to suppress bash-4.4 warnings when reading the
file via bash command substitution.  For some reason, wso2 appends null
characters to the end of the pid file.  And bash-4.4 generates a warning
(simply a message to stderr) when the output of a $() command substitution
contains null characters, resulting in the bash warnings being printed to the
user console when running services-start, services-stop and during upgrades.

The fix is basically to strip out the null characters via "sed -e 's\x0//g'"
instead of just "cat".

This is similar to the fix in wso2server.sh for DEP-583.

Also fixed a separate, unrelated issue in the bin/start check_pid_file()
function (detected while fixiing this bug).  The function was using the
$SL_SERVICES_PID pidfile in the function instead of the $pidfile that was
passed in.  So we could possibly have removed the SL services pidfile in
check_pid_file(), where we really intended to remove the WSO2 pidfile.